Removing bugs by removing code! [fix] removed b8r.removeByPath which was solely used by b8r.removeListInstance, switched to b8r.remove [fix] b8r.remove did not work for id-paths, so fixed that [fix] b8r.removeListInstance did not work if list-binding had no id-path

+++ source/b8r.registry.js
@@ -1279,14 +1279,18 @@ b8r.remove('remove-test')
 ~~~~
 */
 const remove = (path, update = true) => {
-  deleteByPath(registry, path)
-  if (update) {
-    touch(path)
-    /* touch array containing the element if appropriate */
-    ;[, path] = path.match(/^(.+)\[[^\]]+\]$/) || []
-    if (path) {
-      touch(path)
+  const [,listPath] = path.match(/^(.*)\[[^\[]*\]$/) || []
+  if (listPath) {
+    const list = getByPath(registry, listPath)
+    const item = getByPath(registry, path)
+    const index = list.indexOf(item)
+    if (index !== -1) {
+      list.splice(index, 1)
+      if (update) touch(listPath) 
     }
+  } else {
+    deleteByPath(registry, path)
+    if (update) touch(path)
   }
 }
